Move monkey patch to apply, revert functions
