+++ garnett/fields.py
@@ -6,7 +6,7 @@ from django.db.models import CharField, JSONField, TextField
 from django.db.models import F
 from django.db.models.fields.json import KeyTextTransform
 from django.utils.translation import gettext as _
-from dataclasses import dataclass, make_dataclass
+from dataclasses import make_dataclass
 from langcodes import Language
 import logging
 
@@ -238,95 +238,6 @@ FORMFIELD_FOR_DBFIELD_DEFAULTS.update(
         TranslatedTextField: {"widget": widgets.AdminTextareaWidget},
     }
 )
-
-# This is needed to allow values/values_list and F lookups to work
-# The most dangerous monkey patch of all time
-from django.db.models.sql import query
-
-
-@dataclass
-class JoinInfo:
-    final_field: "typing.Any"
-    targets: "typing.Any"
-    opts: "typing.Any"
-    joins: "typing.Any"
-    path: "typing.Any"
-    transform_function_func: "typing.Any"
-
-    @property
-    def transform_function(self):
-        if isinstance(self.final_field, TranslatedFieldBase):
-            # needed for below
-            import functools
-            from django.core.exceptions import FieldError
-
-            # If its a partial, it must have had a transformer applied - leave it alone!
-            if isinstance(self.transform_function_func, functools.partial):
-                # import pdb; pdb.set_trace()
-
-                return self.transform_function_func
-
-            name = get_current_language()
-
-            # Cloned in from django
-            def transform(field, alias, *, name, previous):
-                try:
-                    wrapped = previous(field, alias)
-                    return self.try_transform(wrapped, name)
-                except FieldError:
-                    # FieldError is raised if the transform doesn't exist.
-                    if isinstance(final_field, Field) and last_field_exception:
-                        raise last_field_exception
-                    else:
-                        raise
-
-            # -------------------
-
-            return functools.partial(
-                transform, name=name, previous=self.transform_function_func
-            )
-
-        return self.transform_function_func
-
-    def try_transform(self, lhs, name):
-        # Cloned in from django
-        import difflib
-
-        """
-        Helper method for build_lookup(). Try to fetch and initialize
-        a transform for name parameter from lhs.
-        """
-        transform_class = lhs.get_transform(name)
-        if transform_class:
-            return transform_class(lhs)
-        else:
-            output_field = lhs.output_field.__class__
-            suggested_lookups = difflib.get_close_matches(
-                name, output_field.get_lookups()
-            )
-            if suggested_lookups:
-                suggestion = ", perhaps you meant %s?" % " or ".join(suggested_lookups)
-            else:
-                suggestion = "."
-            raise FieldError(
-                "Unsupported lookup '%s' for %s or join on the field not "
-                "permitted%s" % (name, output_field.__name__, suggestion)
-            )
-
-    def __iter__(self):
-        # Necessary to mimic a tuple
-        for x in [
-            self.final_field,
-            self.targets,
-            self.opts,
-            self.joins,
-            self.path,
-            self.transform_function,
-        ]:
-            yield x
-
-
-query.JoinInfo = JoinInfo
 
 
 # Based on: https://code.djangoproject.com/ticket/29769#comment:5
